Extract operations read model
